feat: Move MediaRepository to new folder structure (#18248)
