+++ src/script/components/calling/CallingCell/CallingCell.test.tsx
@@ -26,9 +26,9 @@ import {CALL_TYPE, STATE as CALL_STATE} from '@wireapp/avs';
 import {Call} from 'Repositories/calling/Call';
 import {CallingRepository} from 'Repositories/calling/CallingRepository';
 import {Participant} from 'Repositories/calling/Participant';
+import {MediaDevicesHandler} from 'Repositories/media/MediaDevicesHandler';
 import {Conversation} from 'src/script/entity/Conversation';
 import {User} from 'src/script/entity/User';
-import {MediaDevicesHandler} from 'src/script/media/MediaDevicesHandler';
 import {TeamState} from 'src/script/team/TeamState';
 import {CallActions} from 'src/script/view_model/CallingViewModel';
 import {createUuid} from 'Util/uuid';
