+++ src/script/components/calling/ChooseScreen.tsx
@@ -22,7 +22,7 @@ import {Fragment, useCallback, useEffect} from 'react';
 import {container} from 'tsyringe';
 
 import {CallState} from 'Repositories/calling/CallState';
-import {ElectronDesktopCapturerSource} from 'src/script/media/MediaDevicesHandler';
+import {ElectronDesktopCapturerSource} from 'Repositories/media/MediaDevicesHandler';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {t} from 'Util/LocalizerUtil';
