+++ src/script/main/app.ts
@@ -61,6 +61,7 @@ import {GiphyRepository} from 'Repositories/giphy/GiphyRepository';
 import {GiphyService} from 'Repositories/giphy/GiphyService';
 import {IntegrationRepository} from 'Repositories/integration/IntegrationRepository';
 import {IntegrationService} from 'Repositories/integration/IntegrationService';
+import {MediaRepository} from 'Repositories/media/MediaRepository';
 import {initializeDataDog} from 'Util/DataDog';
 import {DebugUtil} from 'Util/DebugUtil';
 import {Environment} from 'Util/Environment';
@@ -84,7 +85,6 @@ import {ClientError, CLIENT_ERROR_TYPE} from '../error/ClientError';
 import {TeamError} from '../error/TeamError';
 import {externalUrl} from '../externalRoute';
 import {startNewVersionPolling} from '../lifecycle/newVersionHandler';
-import {MediaRepository} from '../media/MediaRepository';
 import {initMLSGroupConversations, initialiseSelfAndTeamConversations} from '../mls';
 import {joinConversationsAfterMigrationFinalisation} from '../mls/MLSMigration/migrationFinaliser';
 import {NotificationRepository} from '../notification/NotificationRepository';
