+++ src/script/page/MainContent/panels/preferences/avPreferences/MicrophonePreferences.tsx
@@ -20,7 +20,10 @@
 import React, {useEffect, useState} from 'react';
 
 import * as Icon from 'Components/Icon';
-import {MediaDeviceType} from 'src/script/media/MediaDeviceType';
+import {MediaDevicesHandler} from 'Repositories/media/MediaDevicesHandler';
+import {MediaDeviceType} from 'Repositories/media/MediaDeviceType';
+import {MediaStreamHandler} from 'Repositories/media/MediaStreamHandler';
+import {MediaType} from 'Repositories/media/MediaType';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {t} from 'Util/LocalizerUtil';
 import {getLogger} from 'Util/Logger';
@@ -29,9 +32,6 @@ import {DeviceSelect} from './DeviceSelect';
 import {InputLevel} from './InputLevel';
 
 import {Config} from '../../../../../Config';
-import {MediaDevicesHandler} from '../../../../../media/MediaDevicesHandler';
-import {MediaStreamHandler} from '../../../../../media/MediaStreamHandler';
-import {MediaType} from '../../../../../media/MediaType';
 import {PreferencesSection} from '../components/PreferencesSection';
 
 const logger = getLogger('MicrophonePreferences');
