+++ src/script/repositories/calling/CallingRepository.test.ts
@@ -33,9 +33,10 @@ import {Runtime} from '@wireapp/commons';
 import {CallingEvent} from 'Repositories/event/CallingEvent';
 import {CALL} from 'Repositories/event/Client';
 import {EventRepository} from 'Repositories/event/EventRepository';
+import {MediaDevicesHandler} from 'Repositories/media/MediaDevicesHandler';
+import {MediaType} from 'Repositories/media/MediaType';
 import {Conversation} from 'src/script/entity/Conversation';
 import {User} from 'src/script/entity/User';
-import {MediaType} from 'src/script/media/MediaType';
 import {serverTimeHandler} from 'src/script/time/serverTimeHandler';
 import {TestFactory} from 'test/helper/TestFactory';
 import {createUuid} from 'Util/uuid';
@@ -47,7 +48,6 @@ import {CALL_MESSAGE_TYPE} from './enum/CallMessageType';
 import {LEAVE_CALL_REASON} from './enum/LeaveCallReason';
 import {Participant} from './Participant';
 
-import {MediaDevicesHandler} from '../../media/MediaDevicesHandler';
 import {Core} from '../../service/CoreSingleton';
 import {UserRepository} from '../../user/UserRepository';
