+++ src/script/repositories/calling/CallingRepository.ts
@@ -65,6 +65,9 @@ import {CONSENT_TYPE, MessageRepository, MessageSendingOptions} from 'Repositori
 import {CallingEvent} from 'Repositories/event/CallingEvent';
 import {EventRepository} from 'Repositories/event/EventRepository';
 import {EventSource} from 'Repositories/event/EventSource';
+import type {MediaDevicesHandler} from 'Repositories/media/MediaDevicesHandler';
+import type {MediaStreamHandler} from 'Repositories/media/MediaStreamHandler';
+import {MediaType} from 'Repositories/media/MediaType';
 import {flatten} from 'Util/ArrayUtil';
 import {calculateChildWindowPosition} from 'Util/DOM/caculateChildWindowPosition';
 import {isDetachedCallingFeatureEnabled} from 'Util/isDetachedCallingFeatureEnabled';
@@ -87,9 +90,6 @@ import {Config} from '../../Config';
 import {Conversation} from '../../entity/Conversation';
 import type {User} from '../../entity/User';
 import {NoAudioInputError} from '../../error/NoAudioInputError';
-import type {MediaDevicesHandler} from '../../media/MediaDevicesHandler';
-import type {MediaStreamHandler} from '../../media/MediaStreamHandler';
-import {MediaType} from '../../media/MediaType';
 import {APIClient} from '../../service/APIClientSingleton';
 import {Core} from '../../service/CoreSingleton';
 import {TeamState} from '../../team/TeamState';
