+++ src/script/repositories/media/MediaConstraintsHandler.ts
@@ -24,7 +24,7 @@ import {Logger, getLogger} from 'Util/Logger';
 import type {CurrentAvailableDeviceId} from './MediaDevicesHandler';
 import {VIDEO_QUALITY_MODE} from './VideoQualityMode';
 
-import {UserState} from '../user/UserState';
+import {UserState} from '../../user/UserState';
 
 interface Config {
   CONSTRAINTS: {
