+++ src/script/repositories/media/MediaEmbeds.test.ts
@@ -20,8 +20,9 @@
 import type {TypeUtil} from '@wireapp/commons';
 
 import {Configuration, Config} from 'src/script/Config';
-import {MediaEmbeds} from 'src/script/media/MediaEmbeds';
-import {MediaParser} from 'src/script/media/MediaParser';
+
+import {MediaEmbeds} from './MediaEmbeds';
+import {MediaParser} from './MediaParser';
 
 describe('MediaEmbeds', () => {
   // Will test all common link variations
