+++ src/script/repositories/media/MediaRepository.ts
@@ -21,7 +21,7 @@ import {MediaConstraintsHandler} from './MediaConstraintsHandler';
 import {MediaDevicesHandler} from './MediaDevicesHandler';
 import {MediaStreamHandler} from './MediaStreamHandler';
 
-import type {PermissionRepository} from '../permission/PermissionRepository';
+import type {PermissionRepository} from '../../permission/PermissionRepository';
 
 export class MediaRepository {
   public constraintsHandler: MediaConstraintsHandler;
