+++ src/script/repositories/media/MediaStreamError.ts
