+++ src/script/repositories/media/MediaStreamHandler.ts
@@ -29,13 +29,13 @@ import {MEDIA_STREAM_ERROR} from './MediaStreamError';
 import {isMediaStreamReadDeviceError, MEDIA_STREAM_ERROR_TYPES} from './MediaStreamErrorTypes';
 import {MediaType} from './MediaType';
 
-import {MediaError} from '../error/MediaError';
-import {NoAudioInputError} from '../error/NoAudioInputError';
-import {PermissionError} from '../error/PermissionError';
-import type {PermissionRepository} from '../permission/PermissionRepository';
-import {PermissionStatusState} from '../permission/PermissionStatusState';
-import {PermissionType} from '../permission/PermissionType';
-import {Warnings} from '../view_model/WarningsContainer';
+import {MediaError} from '../../error/MediaError';
+import {NoAudioInputError} from '../../error/NoAudioInputError';
+import {PermissionError} from '../../error/PermissionError';
+import type {PermissionRepository} from '../../permission/PermissionRepository';
+import {PermissionStatusState} from '../../permission/PermissionStatusState';
+import {PermissionType} from '../../permission/PermissionType';
+import {Warnings} from '../../view_model/WarningsContainer';
 
 export class MediaStreamHandler {
   static get CONFIG() {
