+++ src/script/view_model/CallingViewModel.mocks.ts
@@ -25,11 +25,11 @@ import {CALL_TYPE, CONV_TYPE} from '@wireapp/avs';
 import {Call} from 'Repositories/calling/Call';
 import {CallingRepository} from 'Repositories/calling/CallingRepository';
 import {CallState} from 'Repositories/calling/CallState';
+import {MediaDevicesHandler} from 'Repositories/media/MediaDevicesHandler';
 
 import {CallingViewModel} from './CallingViewModel';
 
 import {Conversation} from '../entity/Conversation';
-import {MediaDevicesHandler} from '../media/MediaDevicesHandler';
 import {Core} from '../service/CoreSingleton';
 
 export const mockCallingRepository = {
