+++ src/script/view_model/CallingViewModel.ts
@@ -37,6 +37,8 @@ import {CallState, DesktopScreenShareMenu} from 'Repositories/calling/CallState'
 import {LEAVE_CALL_REASON} from 'Repositories/calling/enum/LeaveCallReason';
 import {ConversationState} from 'Repositories/conversation/ConversationState';
 import {ConversationVerificationState} from 'Repositories/conversation/ConversationVerificationState';
+import type {ElectronDesktopCapturerSource, MediaDevicesHandler} from 'Repositories/media/MediaDevicesHandler';
+import type {MediaStreamHandler} from 'Repositories/media/MediaStreamHandler';
 import {replaceLink, t} from 'Util/LocalizerUtil';
 import {matchQualifiedIds} from 'Util/QualifiedId';
 import {safeWindowOpen} from 'Util/SanitizationUtil';
@@ -44,8 +46,6 @@ import {safeWindowOpen} from 'Util/SanitizationUtil';
 import {Config} from '../Config';
 import type {Conversation} from '../entity/Conversation';
 import type {User} from '../entity/User';
-import type {ElectronDesktopCapturerSource, MediaDevicesHandler} from '../media/MediaDevicesHandler';
-import type {MediaStreamHandler} from '../media/MediaStreamHandler';
 import type {PermissionRepository} from '../permission/PermissionRepository';
 import {PermissionStatusState} from '../permission/PermissionStatusState';
 import {PropertiesRepository} from '../properties/PropertiesRepository';
